+++ packages/__cxstarter__/src/styleguide/Page/StyleGuideTemplate/Footer.tsx
@@ -1,10 +1,22 @@
 import React from 'react';
 import { asStyleGuideTemplateToken, cxStyleGuideTemplate } from '@bodiless/cx-templates';
-import { flowHoc, replaceWith } from '@bodiless/fclasses';
+import { flowHoc, on, replaceWith } from '@bodiless/fclasses';
+import { FooterClean, cxFooter } from '@bodiless/cx-layout';
 
-export const Footer = asStyleGuideTemplateToken(cxStyleGuideTemplate.FooterOnly, {
+// Apply similar method as done in the cxLayout token.
+const StickFooterLayout = asStyleGuideTemplateToken({
+  ...cxStyleGuideTemplate.NoLayout,
+  Theme: {
+    ...cxStyleGuideTemplate.NoLayout.Theme,
+    Container: 'flex flex-col h-screen',
+    DescriptionWrapper: 'flex-grow',
+  },
+});
+
+export const Footer = asStyleGuideTemplateToken(StickFooterLayout, {
   Meta: flowHoc.meta.term('Token')('Footer'),
   Content: {
     Title: replaceWith(() => <>Footer</>),
+    Examples: on(FooterClean)(cxFooter.Default),
   },
 });