+++ packages/bodiless-webpack/src/index.ts
@@ -13,4 +13,5 @@
  */
 
 export { addTokenShadowPlugin, createTokenShadowPlugin } from './tokenShadowPlugin';
+export { addStatoscopePlugin, createStatoscopePlugin } from './statoscopePlugin';
 export { addStaticReplacementPlugin, createStaticReplacementPlugin } from './staticReplacementPlugin';