+++ packages/cx-layout/src/components/Layout/tokens/index.ts
@@ -14,6 +14,6 @@
 
 import cxLayout from './cxLayout';
 
-export { MAIN_CONTENT_ID } from './constants';
+export { LayoutIds } from './constants';
 
 export { cxLayout };