+++ packages/cx-templates/src/components/StyleGuideTemplate/tokens/cxStyleGuideTemplate.ts
@@ -12,21 +12,24 @@
  * limitations under the License.
  */
 
-import { cxTypography } from '@bodiless/cx-elements';
+import omit from 'lodash/omit';
+import { cxSpacing, cxTypography } from '@bodiless/cx-elements';
 import {
   EditorPlainClean, cxEditorPlain, RichTextClean, cxRichText
 } from '@bodiless/cx-editors';
 import { withNodeKey } from '@bodiless/core';
-import { on, replaceWith, Fragment } from '@bodiless/fclasses';
+import {
+  on, replaceWith, Fragment, as
+} from '@bodiless/fclasses';
 import { cxLayout, LayoutClean } from '@bodiless/cx-layout';
 import { asStyleGuideTemplateToken } from '../StyleGuideTemplateClean';
 
-const Default = asStyleGuideTemplateToken({
-  Theme: {
-    TitleWrapper: cxTypography.H1,
-  },
+const Base = asStyleGuideTemplateToken({
   Components: {
     Wrapper: on(LayoutClean)(cxLayout.Default),
+  },
+  Theme: {
+    TitleWrapper: cxTypography.H1,
   },
   Editors: {
     Title: on(EditorPlainClean)(cxEditorPlain.Default),
@@ -39,31 +42,29 @@ const Default = asStyleGuideTemplateToken({
   },
 });
 
-const NoLayout = asStyleGuideTemplateToken({
-  ...Default,
+const Default = asStyleGuideTemplateToken({
+  ...Base,
+  Theme: {
+    ...Base.Theme,
+    Container: as(
+      cxSpacing.WithSiteMargin,
+      cxSpacing.WithSiteXLConstraint
+    ),
+  },
+});
+
+const NoLayout = asStyleGuideTemplateToken(omit(Base, 'Components'), {
   Components: {
-    ...Default.Components,
+    ...Base.Components,
     Wrapper: replaceWith(Fragment),
   },
-});
-
-const HeaderOnly = asStyleGuideTemplateToken({
-  ...Default,
-  Components: {
-    Wrapper: on(LayoutClean)(cxLayout.HeaderOnly),
-  },
-});
-
-const FooterOnly = asStyleGuideTemplateToken({
-  ...Default,
-  Components: {
-    Wrapper: on(LayoutClean)(cxLayout.FooterOnly),
+  Layout: {
+    TitleWrapper: 'container mx-auto',
+    DescriptionWrapper: 'container mx-auto',
   },
 });
 
 export default {
   Default,
   NoLayout,
-  HeaderOnly,
-  FooterOnly,
 };